# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "crm",
    "description": "CRM API Docs",
    "version": "0.0.1-SNAPSHOT"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "security": [
    {
      "Authorization": []
    }
  ],
...
  "paths": {
...
    "/api/v1/segments/{id}": {
      "get": {
        "tags": [
          "Segments API"
        ],
        "operationId": "get_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FailureBody"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessBodySegmentDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Segments API"
        ],
        "operationId": "update_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutSegmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FailureBody"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessBodyPostSegmentUseCaseOut"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Segments API"
        ],
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FailureBody"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/segments": {
      "get": {
        "tags": [
          "Segments API"
        ],
        "operationId": "list_1",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FailureBody"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessBodyListSegmentDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Segments API"
        ],
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostSegmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FailureBody"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessBodyPostSegmentUseCaseOut"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FailureBody": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "JsonNode": {
        "type": "object"
      },
      "PutSegmentRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "conditions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SegmentConditionRequest"
            }
          }
        }
      },
      "SegmentConditionRequest": {
        "required": [
          "field",
          "operator",
          "value",
          "valueType"
        ],
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "operator": {
            "type": "string"
          },
          "valueType": {
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/JsonNode"
          }
        }
      },
      "PostSegmentUseCaseOut": {
        "required": [
          "segment"
        ],
        "type": "object",
        "properties": {
          "segment": {
            "$ref": "#/components/schemas/SegmentDto"
          }
        }
      },
      "SegmentConditionDto": {
        "required": [
          "field",
          "operator",
          "position",
          "value",
          "valueType"
        ],
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "operator": {
            "type": "string"
          },
          "valueType": {
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/JsonNode"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SegmentDto": {
        "required": [
          "active",
          "conditions",
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "conditions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SegmentConditionDto"
            }
          },
          "createdAt": {
            "type": "string"
          }
        }
      },
      "SuccessBodyPostSegmentUseCaseOut": {
        "required": [
          "data",
          "message"
        ],
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PostSegmentUseCaseOut"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "PostSegmentRequest": {
        "required": [
          "conditions",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "conditions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SegmentConditionRequest"
            }
          }
        }
      },
...
      "SuccessBodyListSegmentDto": {
        "required": [
          "data",
          "message"
        ],
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SegmentDto"
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
      "SuccessBodySegmentDto": {
        "required": [
          "data",
          "message"
        ],
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SegmentDto"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "Success": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Authorization": {
        "type": "apiKey",
        "description": "Access Token",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}